[DROOLS-3570] Included PR comments
